# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Belgü",
    "version": "0.5.0"
  },
  "paths": {
...
    "/api/settings": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Read Settings",
        "operationId": "read_settings_api_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationView"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "settings"
        ],
        "summary": "Save Settings",
        "operationId": "save_settings_api_settings_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntegrationPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/llm/models": {
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Available Models",
        "operationId": "available_models_api_settings_llm_models_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LlmCheck"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailableModels"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/llm/test": {
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Test Llm",
        "operationId": "test_llm_api_settings_llm_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LlmCheck"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/sources/test": {
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Test Source",
        "operationId": "test_source_api_settings_sources_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceCheck"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AvailableModels": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "error"
            ],
            "title": "Status"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ModelInfo"
            },
            "type": "array",
            "title": "Items"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "AvailableModels"
      },
...
      "ConnectionResult": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "error"
            ],
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "elapsed_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsed Ms"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "title": "ConnectionResult"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IntegrationPatch": {
        "properties": {
          "active_llm": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "local",
                  "anthropic",
                  "openai",
                  "gemini",
                  "openrouter"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Llm"
          },
          "llm": {
            "additionalProperties": {
              "$ref": "#/components/schemas/LlmDraft"
            },
            "type": "object",
            "title": "Llm"
          },
          "sources": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SourceDraft"
            },
            "type": "object",
            "title": "Sources"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "IntegrationPatch"
      },
      "IntegrationView": {
        "properties": {
          "active_llm": {
            "type": "string",
            "enum": [
              "local",
              "anthropic",
              "openai",
              "gemini",
              "openrouter"
            ],
            "title": "Active Llm"
          },
          "llm": {
            "additionalProperties": {
              "$ref": "#/components/schemas/LlmView"
            },
            "type": "object",
            "title": "Llm"
          },
          "sources": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SourceView"
            },
            "type": "object",
            "title": "Sources"
          },
          "read_only": {
            "type": "boolean",
            "title": "Read Only"
          }
        },
        "type": "object",
        "required": [
          "active_llm",
          "llm",
          "sources",
          "read_only"
        ],
        "title": "IntegrationView"
      },
...
      "LlmCheck": {
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "local",
              "anthropic",
              "openai",
              "gemini",
              "openrouter"
            ],
            "title": "Provider"
          },
          "settings": {
            "$ref": "#/components/schemas/LlmDraft"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "provider"
        ],
        "title": "LlmCheck"
      },
      "LlmDraft": {
        "properties": {
          "api_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 8192
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "clear_key": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clear Key"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "timeout_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 3600.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Seconds"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "backend": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "sglang",
                  "llama.cpp",
                  "openai_compatible"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Backend"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "LlmDraft"
      },
      "LlmView": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "docs_url": {
            "type": "string",
            "title": "Docs Url"
          },
          "base_url": {
            "type": "string",
            "title": "Base Url"
          },
          "backend": {
            "type": "string",
            "title": "Backend"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "timeout_seconds": {
            "type": "integer",
            "title": "Timeout Seconds"
          },
          "key_configured": {
            "type": "boolean",
            "title": "Key Configured"
          }
        },
        "type": "object",
        "required": [
          "label",
          "docs_url",
          "base_url",
          "backend",
          "model_id",
          "timeout_seconds",
          "key_configured"
        ],
        "title": "LlmView"
      },
      "ModelInfo": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "runtime_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runtime Version"
          },
          "text_supported": {
            "type": "boolean",
            "title": "Text Supported",
            "default": true
          },
          "vision_supported": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vision Supported"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id"
        ],
        "title": "ModelInfo"
      },
...
      "SourceCheck": {
        "properties": {
          "provider": {
            "type": "string",
            "maxLength": 40,
            "title": "Provider"
          },
          "settings": {
            "$ref": "#/components/schemas/SourceDraft"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "provider"
        ],
        "title": "SourceCheck"
      },
      "SourceDraft": {
        "properties": {
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 8192
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "clear_key": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clear Key"
          },
          "file_path": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "File Path"
          },
          "mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "api",
                  "local"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SourceDraft"
      },
      "SourceView": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "group": {
            "type": "string",
            "title": "Group"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "docs_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Docs Url"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "accepts_key": {
            "type": "boolean",
            "title": "Accepts Key"
          },
          "key_configured": {
            "type": "boolean",
            "title": "Key Configured"
          },
          "file_path": {
            "type": "string",
            "title": "File Path"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          }
        },
        "type": "object",
        "required": [
          "label",
          "group",
          "description",
          "docs_url",
          "enabled",
          "accepts_key",
          "key_configured",
          "file_path",
          "mode"
        ],
        "title": "SourceView"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}